refactor: Realiza alterações em nomes da aplicação
