+++ src/main/java/one/digitalinnovation/gof/repositories/EnderecoRepository.java
@@ -1,5 +1,6 @@
-package one.digitalinnovation.gof.model;
+package one.digitalinnovation.gof.repositories;
 
+import one.digitalinnovation.gof.model.Endereco;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
